feat(auth): integrate MailConfigService for dynamic frontend URL and update JWT secret handling refactor(auth): adjust imports for password hashing utility functions
The MailConfigService is now injected into the AuthController to dynamically retrieve the frontend URL for redirection after login. The JWT secret is now configured using the AuthConfig module, allowing for better management of configuration values. Additionally, the password hashing utility functions are imported from a shared location to promote code reuse and maintainability.
