+++ src/auth/discord-auth.service.ts
@@ -6,7 +6,7 @@ import {
 import { IAuthProvider } from "../global/interfaces/oauth.interface";
 import { PrismaService } from "../prisma/prisma.service";
 import { DiscordUser } from "../global/types/auth.types";
-import { generatePasswordHash } from "../utils/auth";
+import { generatePasswordHash } from "../../shared/auth/auth";
 
 import { OAuthConfig } from "src/config/0auth/oauthConfig.interface";
 @Injectable()
